+++ __tests__/__system__/cli/delete/data-set/__scripts__/command/command_delete_data_set.sh
@@ -2,4 +2,4 @@
 dsn=$1
 set -e
 
-bright zos-ftp delete data-set "$dsn" -f
+zowe zos-ftp delete data-set "$dsn" -f