+++ __tests__/__system__/cli/delete/uss-file/__scripts__/command/command_delete_uss_file.sh
@@ -2,4 +2,4 @@
 ussfile=$1
 set -e
 
-bright zos-ftp delete uss-file $ussfile -f
+zowe zos-ftp delete uss-file $ussfile -f