+++ __tests__/__system__/cli/list/data-set-members/__scripts__/command/command_list_data_set_members.sh
@@ -2,4 +2,4 @@
 dsn=$1
 set -e
 
-bright zos-ftp list data-set-members "$dsn"
+zowe zos-ftp list data-set-members "$dsn"