+++ __tests__/__system__/cli/list/job/__scripts__/command/command_list_job.sh
@@ -3,4 +3,4 @@ pre=$1
 owner=$2
 set -e
 
-bright zos-ftp list jobs --prefix $pre --owner $owner
+zowe zos-ftp list jobs --prefix $pre --owner $owner