+++ __tests__/__system__/cli/list/job/__scripts__/command/command_list_job_noprefix.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 set -e
 
-bright zos-ftp list jobs
+zowe zos-ftp list jobs