+++ __tests__/__system__/cli/submit/local-file/__scripts__/submit_local_file_help.sh
@@ -2,7 +2,7 @@
 set -e
 
 echo "================z/OS VIEW SUBMIT LOCAL-FILE HELP==============="
-bright zos-ftp submit local-file --help
+zowe zos-ftp submit local-file --help
 
 echo "================z/OS FTP SUBMIT LOCAL-FILE HELP WITH RFJ==========="
-bright zos-ftp submit local-file --help --rfj
+zowe zos-ftp submit local-file --help --rfj