+++ __tests__/__system__/cli/upload/file-to-uss-file/__scripts__/upload_file_to_uss_file_help.sh
@@ -2,7 +2,7 @@
 set -e
 
 echo "================z/OS FTP UPLOAD FILE-TO-USS-FILE HELP ==============="
-bright zos-ftp upload file-to-uss-file --help
+zowe zos-ftp upload file-to-uss-file --help
 
 echo "================z/OS FTP UPLOAD FILE-TO-USS-FILE HELP WITH RFJ==========="
-bright zos-ftp ul ftuf --help --rfj
+zowe zos-ftp ul ftuf --help --rfj