+++ __tests__/__system__/cli/view/all-spool-by-jobid/__scripts__/command/command_view_all_spool_by_jobid.sh
@@ -2,4 +2,4 @@
 jobid=$1
 set -e
 
-bright zos-ftp view all-spool-by-jobid "$jobid"
+zowe zos-ftp view all-spool-by-jobid "$jobid"