+++ __tests__/__system__/cli/view/spool-file-by-id/__scripts__/view_spool_file_by_id_help.sh
@@ -2,7 +2,7 @@
 set -e
 
 echo "================z/OS FTP VIEW SPOOL-FILE-BY-ID HELP==============="
-bright zos-ftp view spool-file-by-id --help
+zowe zos-ftp view spool-file-by-id --help
 
 echo "================z/OS FTP VIEW SPOOL-FILE-BY-ID HELP WITH RFJ==========="
-bright zos-ftp vw sfbi --help --rfj
+zowe zos-ftp vw sfbi --help --rfj